Refactor: Move Save/Unsave button functionality to Media
- Moved the code that was previously inside the Library app into Media, making
the buttons part of the functionality of the detail pages already in place.
- Deleted the 'library' app,
- Fixed typos in the SavedAnime and SavedManga classes in admin.
