+++ media/templates/media/anime/detail.html
@@ -5,6 +5,6 @@
         <h1>{{ anime.title }}</h1>
         <img class="w-64 object-cover rounded-lg" src="{{ anime.cover_image }}" alt="{{ anime.title }}">
         {{ anime.country_of_origin }}
-        {% include "library/includes/save_button.html" with content_type="anime" object_id=anime.pk %}
+        {% include "media/save_button.html" with content_type="anime" object_id=anime.pk %}
     </div>    
 {% endblock %}
