+++ media/templates/media/movie/detail.html
@@ -6,6 +6,6 @@
         <img class="w-64 object-cover rounded-lg" src="https://image.tmdb.org/t/p/w342{{ movie.poster_path }}" alt="{{ movie.title }}">
         {{ movie.overview | linebreaks }}
         {{ movie.release_date }}
-        {% include "library/includes/save_button.html" with content_type="movie" object_id=movie.pk %}
+        {% include "media/save_button.html" with content_type="movie" object_id=movie.pk %}
     </div>
 {% endblock %}
