+++ media/templates/media/save_button.html
@@ -1,6 +1,6 @@
 {% if user.is_authenticated %}
   {% if is_saved %}
-    <form method="post" action="{% url 'library:unsave' content_type object_id %}">
+    <form method="post" action="{% url 'media:unsave' content_type object_id %}">
       {% csrf_token %}
       <input type="hidden" name="next" value="{{ request.path }}">
       <button type="submit"
@@ -9,7 +9,7 @@
       </button>
     </form>
   {% else %}
-    <form method="post" action="{% url 'library:save' content_type object_id %}">
+    <form method="post" action="{% url 'media:save' content_type object_id %}">
       {% csrf_token %}
       <input type="hidden" name="next" value="{{ request.path }}">
       <button type="submit"
