+++ mysite/settings.py
@@ -47,7 +47,6 @@ INSTALLED_APPS = [
     "media",
     "tailwind",
     "theme",
-    "library",
 ]
 
 TAILWIND_APP_NAME = "theme"
