+++ mysite/urls.py
@@ -23,7 +23,6 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("", include("media.urls")),
     path("accounts/", include("account.urls")),
-    path("library/", include("library.urls")),
 ]
 
 if settings.DEBUG:
